Fixed pr issues
