+++ forui/lib/src/widgets/avatar/avatar.dart
@@ -5,6 +5,8 @@ import 'package:meta/meta.dart';
 
 import 'package:forui/forui.dart';
 
+part 'avatar_content.dart';
+
 /// An image element with a fallback for representing the user.
 ///
 /// use image property to provide a profile image displayed within the circle.
@@ -31,22 +33,22 @@ class FAvatar extends StatelessWidget {
   /// Creates an [FAvatar].
   FAvatar({
     required ImageProvider image,
-    Widget? placeholder,
-    this.size = 40.0,
     this.style,
+    this.size = 40.0,
+    Widget? placeholder,
     super.key,
   }) : placeholderBuilder = ((context, style) => _AvatarContent(
+              image: image,
               style: style,
               size: size,
-              image: image,
               placeholder: placeholder,
             ));
 
   /// Creates a [FAvatar] with custom child.
   FAvatar.raw({
     Widget? child,
-    this.size = 40.0,
     this.style,
+    this.size = 40.0,
     super.key,
   }) : placeholderBuilder = ((context, style) =>
             child ??
@@ -167,91 +169,4 @@ final class FAvatarStyle with Diagnosticable {
 
   @override
   int get hashCode => backgroundColor.hashCode ^ foregroundColor.hashCode ^ fadeInDuration.hashCode ^ text.hashCode;
-}
-
-class _AvatarContent extends StatelessWidget {
-  final FAvatarStyle? style;
-
-  /// The circle's size.
-  final double size;
-
-  /// The profile image displayed within the circle.
-  ///
-  /// If the user's initials are used, use [placeholder] instead.
-  final ImageProvider image;
-
-  /// The fallback widget displayed if [image] fails to load.
-  ///
-  /// Typically used to display the user's initials using a [Text] widget
-  /// styled with [FAvatarStyle.backgroundColor].
-  final Widget? placeholder;
-
-  const _AvatarContent({
-    required this.style,
-    required this.size,
-    required this.image,
-    this.placeholder,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    final style = this.style ?? context.theme.avatarStyle;
-
-    final placeholder = this.placeholder ?? _IconPlaceholder(style: style, size: size);
-
-    return Image(
-      filterQuality: FilterQuality.medium,
-      image: image,
-      errorBuilder: (context, exception, stacktrace) => placeholder,
-      frameBuilder: (context, child, frame, wasSynchronouslyLoaded) {
-        if (wasSynchronouslyLoaded) {
-          return child;
-        }
-        return AnimatedSwitcher(
-          duration: const Duration(milliseconds: 500),
-          child: frame == null ? placeholder : child,
-        );
-      },
-      loadingBuilder: (context, child, loadingProgress) {
-        if (loadingProgress == null) {
-          return child;
-        }
-        return placeholder;
-      },
-      fit: BoxFit.cover,
-    );
-  }
-
-  @override
-  void debugFillProperties(DiagnosticPropertiesBuilder properties) {
-    super.debugFillProperties(properties);
-    properties
-      ..add(DiagnosticsProperty('style', style))
-      ..add(DoubleProperty('size', size))
-      ..add(DiagnosticsProperty('image', image));
-  }
-}
-
-class _IconPlaceholder extends StatelessWidget {
-  final double size;
-  final FAvatarStyle? style;
-
-  const _IconPlaceholder({required this.size, this.style});
-
-  @override
-  Widget build(BuildContext context) {
-    final style = this.style ?? context.theme.avatarStyle;
-    return FAssets.icons.userRound(
-      height: size / 2,
-      colorFilter: ColorFilter.mode(style.foregroundColor, BlendMode.srcIn),
-    );
-  }
-
-  @override
-  void debugFillProperties(DiagnosticPropertiesBuilder properties) {
-    super.debugFillProperties(properties);
-    properties
-      ..add(DoubleProperty('size', size))
-      ..add(DiagnosticsProperty('style', style));
-  }
 }
