+++ forui/lib/widgets.dart
@@ -19,7 +19,7 @@ library forui.widgets;
 import 'package:forui/forui.dart';
 
 export 'src/widgets/alert/alert.dart' hide Variant;
-export 'src/widgets/avatar.dart';
+export 'src/widgets/avatar/avatar.dart';
 export 'src/widgets/badge/badge.dart' hide Variant;
 export 'src/widgets/bottom_navigation_bar/bottom_navigation_bar.dart';
 export 'src/widgets/button/button.dart' hide Variant;
